requirements.txt
numpy~=1.23.*
scikit-learn~=1.2.1
tqdm>=4.41.0
imbalanced-learn~=0.10.1
Pillow~=9.2.0
scipy~=1.10.*

plotly~=5.15.0
kaleido~=0.2.1
pandas~=2.0.3

sounddevice~=0.4.6

tensorflow-gpu==2.5.0
tensorflow-addons~=0.13.0
tensorflow-model-optimization==0.7.0

pyflann-py3

tornado==6.1